찜목록 관련 API 추가

+++ HappyHouse_Final/src/main/java/com/ssafy/happyhouse/model/UserDto.java
@@ -18,7 +18,7 @@ public class UserDto {
 	@ApiModelProperty(value = "찜한 목록의 평균 가격", example = "1")
 	private int avg_price;
 	@ApiModelProperty(value = "찜한 목록의 평균 크기", example = "12.32")
-	private double avg_size;
+	private int avg_size;
 	@ApiModelProperty(value = "찜한 목록의 평균 층수", example = "1")
 	private int avg_floor;
 	@ApiModelProperty(value = "가입 날짜")
@@ -60,10 +60,10 @@ public class UserDto {
 	public void setAvg_price(int avg_price) {
 		this.avg_price = avg_price;
 	}
-	public double getAvg_size() {
+	public int getAvg_size() {
 		return avg_size;
 	}
-	public void setAvg_size(double avg_size) {
+	public void setAvg_size(int avg_size) {
 		this.avg_size = avg_size;
 	}
 	public int getAvg_floor() {
